update octopus web to new interface
