+++ binding/web/.gitignore
@@ -1,4 +1,5 @@
 node_modules
-octopus-web-*
-*.log
-language/**/*.ts
+dist
+lib/pv_octopus*.wasm
+test/octopus_params.js
+test/octopus_params.pv
