+++ binding/web/.npmignore
@@ -1,0 +1,5 @@
+.idea
+node_modules
+.DS_Store
+test
+lib
