+++ binding/web/.prettierignore
@@ -1,0 +1,2 @@
+.eslintrc.js
+src/wasm/pv_octopus*
