--- binding/web/core/.npmignore
@@ -1,3 +1,0 @@
-.idea
-node_modules
-.DS_Store
