--- binding/web/core/README.md
@@ -1,14 +1,0 @@
-# Octopus-web-core
-
-Package containing types for Octopus web bindings.
-
-## Build 
-
-To build Octopus typescript types, run:
-
-```console
-yarn
-yarn build
-```
-
-This will build the types required for Octopus web and web-framework bindings.
