--- binding/web/core/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@picovoice/octopus-web-core",
-  "version": "1.0.0",
-  "description": "Octopus types for web bindings",
-  "main": "src/index.ts",
-  "types": "dist/types/index.d.ts",
-  "author": "Picovoice",
-  "license": "Apache-2.0",
-  "scripts": {
-    "build:types": "tsc --declaration --declarationMap --emitDeclarationOnly --outDir ./dist/types",
-    "build": "npm-run-all --parallel build:**"
-  },
-  "devDependencies": {
-    "npm-run-all": "^4.1.5",
-    "typescript": "^4.5.4"
-  }
-}
