+++ binding/web/module.d.ts
@@ -1,0 +1,9 @@
+declare module "*.wasm" {
+  const content: string;
+  export default content;
+}
+
+declare module 'web-worker:*' {
+  const WorkerFactory: new () => Worker;
+  export default WorkerFactory;
+}
