+++ binding/web/package.json
@@ -1,24 +1,64 @@
 {
-  "name": "octopus-web-template",
-  "private": true,
-  "version": "1.1.0",
-  "type": "module",
-  "description": "Template project for building octopus-web-* packages",
-  "main": ".eslintrc.js",
-  "scripts": {
-    "base64": "node scripts/base64.js",
-    "build": "yarn base64 && yarn template && yarn copytests",
-    "template": "node scripts/template.js",
-    "copytests": "node scripts/copy_tests.js"
-  },
-  "keywords": [],
+  "name": "@picovoice/octopus-web",
+  "description": "Octopus library for web browsers (via WebAssembly)",
   "author": "Picovoice Inc",
   "license": "Apache-2.0",
+  "version": "1.1.7",
+  "keywords": [
+    "octopus",
+    "web",
+    "voice",
+    "speech",
+    "recognition",
+    "ai",
+    "speech-to-index",
+    "offline"
+  ],
+  "entry": "src/index.ts",
+  "module": "dist/esm/index.js",
+  "iife": "dist/iife/index.js",
+  "types": "dist/types/index.d.ts",
+  "scripts": {
+    "build:all": "rollup --config",
+    "build:types": "tsc --declaration --declarationMap --emitDeclarationOnly --outDir ./dist/types",
+    "build": "npm-run-all --parallel build:**",
+    "lint": "eslint . --ext .js,.ts",
+    "prepack": "npm-run-all build",
+    "start": "cross-env TARGET='debug' rollup --config --watch",
+    "watch": "rollup --config --watch",
+    "format": "prettier --write \"**/*.{js,ts,json}\"",
+    "postinstall": "yarn copywasm",
+    "copywasm": "node scripts/copy_wasm.js",
+    "setup-test": "cp ./lib/octopus_params.pv ./test && npx pvbase64 -i ./test/octopus_params.pv -o ./test/octopus_params.js"
+  },
+  "dependencies": {
+    "@picovoice/web-utils": "=1.2.1"
+  },
   "devDependencies": {
-    "edit-json-file": "^1.5.0",
-    "ncp": "^2.0.0"
+    "@babel/core": "^7.18.2",
+    "@babel/plugin-transform-runtime": "^7.18.2",
+    "@babel/preset-env": "^7.18.2",
+    "@babel/runtime": "^7.18.3",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^18.1.0",
+    "@rollup/plugin-node-resolve": "^11.2.1",
+    "@rollup/pluginutils": "^4.2.1",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
+    "async-mutex": "^0.3.2",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.13.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.6.2",
+    "rollup": "^2.70.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "^0.31.2",
+    "rollup-plugin-web-worker-loader": "^1.6.1",
+    "tslib": "^2.3.1",
+    "typescript": "^4.6.3",
+    "wasm-feature-detect": "^1.2.11"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   }
 }
