+++ binding/web/scripts/copy_wasm.js
@@ -1,0 +1,28 @@
+const fs = require("fs");
+const { join } = require("path");
+
+const wasmFiles = ["pv_octopus.wasm", "pv_octopus_simd.wasm"]
+
+console.log("Copying the WASM model...");
+
+const sourceDirectory = join(
+  __dirname,
+  "..",
+  "..",
+  "..",
+  "lib",
+  "wasm"
+);
+
+const outputDirectory = join(__dirname, "..", "lib");
+
+try {
+  fs.mkdirSync(outputDirectory, { recursive: true });
+  wasmFiles.forEach(file => {
+    fs.copyFileSync(join(sourceDirectory, file), join(outputDirectory, file))
+  })
+} catch (error) {
+  console.error(error);
+}
+
+console.log("... Done!");
