--- binding/web/template/.eslintignore
@@ -1,5 +1,0 @@
-node_modules
-dist
-*.worker.js
-**/rollup.config.js
-.eslintrc.js
