--- binding/web/template/.npmignore
@@ -1,4 +1,0 @@
-.idea
-node_modules
-.DS_Store
-test
