--- binding/web/template/.prettierignore
@@ -1,2 +1,0 @@
-.eslintrc.js
-src/wasm/pv_octopus*
