--- binding/web/template/.prettierrc
@@ -1,8 +1,0 @@
-{
-  "semi": true,
-  "trailingComma": "es5",
-  "singleQuote": true,
-  "printWidth": 80,
-  "tabWidth": 2,
-  "arrowParens": "avoid"
-}
