--- binding/web/template/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@picovoice/octopus-web-worker",
-  "version": "1.1.6",
-  "description": "Octopus library for web browsers (via WebAssembly)",
-  "author": "Picovoice Inc",
-  "license": "Apache-2.0",
-  "entry": "src/index.ts",
-  "module": "dist/esm/index.js",
-  "iife": "dist/iife/index.js",
-  "types": "dist/types/index.d.ts",
-  "keywords": [
-    "octopus",
-    "web",
-    "voice",
-    "speech",
-    "recognition",
-    "ai",
-    "speech-to-index",
-    "offline"
-  ],
-  "scripts": {
-    "build:all": "rollup --config",
-    "build:types": "tsc --declaration --declarationMap --emitDeclarationOnly --outDir ./dist/types",
-    "build": "npm-run-all --parallel build:**",
-    "lint": "eslint . --ext .js,.ts",
-    "prepack": "npm-run-all build",
-    "start": "cross-env TARGET='debug' rollup --config --watch",
-    "watch": "rollup --config --watch",
-    "format": "prettier --write \"**/*.{js,ts,json}\""
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.13",
-    "@babel/plugin-transform-runtime": "^7.12.15",
-    "@babel/preset-env": "^7.12.13",
-    "@babel/runtime": "^7.12.13",
-    "@picovoice/web-utils": "^1.0.1",
-    "@rollup/plugin-babel": "^5.2.3",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.1.1",
-    "@typescript-eslint/eslint-plugin": "^4.15.0",
-    "@typescript-eslint/parser": "^4.15.0",
-    "async-mutex": "^0.3.2",
-    "cross-env": "^7.0.3",
-    "eslint": "^7.19.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.2.1",
-    "rollup": "^2.38.5",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.29.0",
-    "rollup-plugin-web-worker-loader": "^1.6.0",
-    "tslib": "^2.1.0",
-    "typescript": "^4.1.5"
-  },
-  "dependencies": {
-    "@picovoice/octopus-web-core": "^1.0.0"
-  }
-}
