--- binding/web/template/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": false,
-    "allowSyntheticDefaultImports": true,
-    "downlevelIteration": true,
-    "isolatedModules": false,
-    "lib": ["esnext", "dom"],
-    "module": "esnext",
-    "moduleResolution": "node",
-    "noEmit": false,
-    "outDir": "./dist",
-    "removeComments": false,
-    "resolveJsonModule": true,
-    "sourceMap": true,
-    "strict": true,
-    "target": "esnext"
-  },
-  "exclude": ["node_modules", "dist"]
-}
