--- binding/web/template/workers.d.ts
@@ -1,4 +1,0 @@
-declare module 'web-worker:*' {
-  const WorkerFactory: new () => Worker;
-  export default WorkerFactory;
-}
