+++ demo/web/.gitignore
@@ -4,3 +4,4 @@ node_modules
 dist/
 *.log
 .DS_Store
+octopus_params.js
