+++ demo/web/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "private": true,
   "scripts": {
-    "start": "yarn run serve"
+    "postinstall": "npx pvbase64 -i ../../lib/common/octopus_params.pv -o ./octopus_params.js",
+    "start": "yarn run http-server -a 0.0.0.0 -p 5000"
   },
   "keywords": [
     "Picovoice",
@@ -17,9 +18,9 @@
   "author": "Picovoice Inc",
   "license": "Apache-2.0",
   "dependencies": {
-    "@picovoice/octopus-web-en-worker": "^1.1.6"
+    "@picovoice/octopus-web-en-worker": "../../binding/web"
   },
   "devDependencies": {
-    "serve": "^11.3.2"
+    "http-server": "^14.0.0"
   }
 }
